requirements.txt
os
gradio
python-dotenv
langchain_groq
langchain_community
langchain_core